cleaned up Arbiter
