+++ bin/ping_network.sh
